Made a sphere and a header for domegame
