+++ Scene.cpp
@@ -4,7 +4,7 @@
 // Constructor
 Scene::Scene(){
 	std::cout << "scene gjord" << std::endl;
-    background = new Quad("avatar3.png", 1.0f, 1.0f);
+    background = new Quad("avatar3", 1.0f, 1.0f);
 }
