+++ domegame.cpp
@@ -1,26 +1,4 @@
-//#include "Player.hpp"
-#include "./Scene.hpp"
-#include <stdlib.h>
-#include <stdio.h>
-#include <cstdlib>
-#include <ctime>
-#ifndef DOMEGAME
-#define DOMEGAME
-
-class DomeGame{
-public:
-	// Constructor
-	DomeGame(sgct::Engine * gEngine);
-
-	void init();
-	void draw() const;
-	size_t textureHandle;
-	Scene * myScene;
-private:
-	
-};
-
-#endif // DOMEGAME
+#include "./DomeGame.hpp"
 
 DomeGame::DomeGame(sgct::Engine * gEngine) {		//Constructor
 	myScene = new Scene();
@@ -34,7 +12,8 @@ void DomeGame::draw() const{
 void DomeGame::init() {
 	sgct::TextureManager::instance()->setAnisotropicFilterSize(8.0f);	
 	sgct::TextureManager::instance()->setCompression(sgct::TextureManager::S3TC_DXT);
-	sgct::TextureManager::instance()->loadTexture("avatar3", "./avatar3.png", true);
+	sgct::TextureManager::instance()->loadTexture("avatar3", "../avatar3.png", true);
+	// = new sgct_utils::SGCTBox(2.0f, sgct_utils::SGCTBox::CubeMap);
 
 	// Enable some openGL stuff
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
